third_party/blink/web_tests/platform/generic/fast/events/scroll-after-click-on-tab-index-expected.txt

ALERT: Scroll position 0
Click the text below and then try to scroll.

Scroll position is more than 0
PASS document.scrollingElement.scrollTop > 0 became true
PASS successfullyParsed is true

TEST COMPLETE











My span
